feat(oracle_rewrite): Organization stuff
