+++ crates/accounts/src/oracle/oracle.masm
@@ -1,0 +1,43 @@
+use.std::sys
+use.miden::account
+
+#! Writes new entry into the oracle's data slots.
+#!
+#! Inputs:  [WORD_1, WORD_2, WORD_3, WORD_4]
+#! Outputs: []
+#!
+export.publish_entry
+    push.0
+    exec.account::set_item
+    dropw dropw
+    # => [WORD_2, WORD_3, WORD_4]
+
+    push.1
+    exec.account::set_item
+    dropw dropw
+    # => [WORD_3, WORD_4]
+
+    push.2
+    exec.account::set_item
+    dropw dropw
+    # => [WORD_4]
+
+    push.3
+    exec.account::set_item
+    dropw dropw
+    # => []
+end
+
+#! Gets entry from the oracle's data slots.
+#!
+#! Inputs:  [storage_slot]
+#! Outputs: [WORD]
+#!
+export.get_entry
+    # make this foreign procedure unique to make sure that we invoke the procedure of the 
+    # foreign account, not the native one
+    push.1 drop
+    exec.account::get_item
+
+    exec.sys::truncate_stack
+end
