+++ crates/accounts/src/publisher/mod.rs
@@ -1,0 +1,1 @@
+pub const PUBLISH_CALL_MASM: &str = include_str!("publish_call.masm");
