+++ crates/types/src/lib.rs
@@ -1,3 +1,7 @@
 pub mod currency;
 pub mod entry;
 pub mod pair;
+
+pub use currency::*;
+pub use entry::*;
+pub use pair::*;
